// models/DynamicTableColumn.tsx
/********************************************************************************
 * Copyright (c) 2021,2022 FEV Consulting GmbH
 * Copyright (c) 2021,2022 Contributors to the CatenaX (ng) GitHub Organisation
 *
 * See the NOTICE file(s) distributed with this work for additional
 * information regarding copyright ownership.
 *
 * This program and the accompanying materials are made available under the
 * terms of the Apache License, Version 2.0 which is available at
 * https://www.apache.org/licenses/LICENSE-2.0.
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 * License for the specific language governing permissions and limitations
 * under the License.
 *
 * SPDX-License-Identifier: Apache-2.0
 ********************************************************************************/

export interface DynamicTableColumn {
  field: string;
  headerName: string;
  editable: boolean;
  sortable: boolean;
  // eslint-disable-next-line
  renderHeader?: () => {};
  // eslint-disable-next-line
  renderCell?: (params: any) => {};
  flex: number;
  headerAlign: 'center' | 'right' | 'left';
  type?: 'singleSelect' | 'date';
  valueOptions?: { value: string; label: string }[];
}